# ActivatorUtilities.CreateInstance: allocation optimization