Some optimizations to try to fix #693.

Before:
```
|          Method |        Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|---------------- |------------:|----------:|----------:|-------------:|-------:|----------:|
|  CreateInstance | 1,317.58 ns | 5.6777 ns | 4.7411 ns |    758,965.3 | 0.0267 |     784 B |
```

After:
```
|          Method |        Mean |      Error |     StdDev |         Op/s |  Gen 0 | Allocated |
|---------------- |------------:|-----------:|-----------:|-------------:|-------:|----------:|
|  CreateInstance | 1,275.97 ns | 26.4718 ns | 44.9511 ns |    783,719.0 | 0.0134 |     448 B |
```